src/requirements.txt
django
celery
djangorestframework
django-celery
django-celery-beat
django-celery-results
psycopg2-binary
requests
gunicorn
django-session-timeout
gevent
django-cors-headers
drf-yasg
geoip2
django-rest-durin
django-db-geventpool
emoji-country-flag
pytz
sentry-sdk
boto3
django-storages
